// src/app/services/models/User.ts
export interface User {
    id:                   number;
    username:             string;
    firstName:            string;
    lastName:             string;
    email:                string;
    phoneNumber:          string;
    photoUrl:             string;
    rating:               number;
    tasks:                string;
    description:          string;
    addressLong:          number;
    addressLarg:          number;
    verifed:              boolean;
    userRoles:            string[];
    cinFileUrl:           string;
    mfFileUrl:            string;

  
  
}